# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Gojek Sentiment API","description":"API analisis sentimen & topik keluhan ulasan Gojek.\n\nEnam endpoint membaca agregat pre-scoring dari PostgreSQL. `/api/predict` memuat model dan berfungsi sebagai demo kapabilitas — bukan sumber data dashboard.","version":"1.0.0"},"paths":{"/api/kpi":{"get":{"tags":["agregat"],"summary":"Ringkasan KPI untuk kartu atas dashboard","operationId":"baca_kpi_api_kpi_get","parameters":[{"name":"min_reviews","in":"query","required":false,"schema":{"type":"integer","minimum":100,"description":"Ambang jumlah ulasan per versi untuk menghitung `jumlah_versi`.","title":"Min Reviews"},"description":"Ambang jumlah ulasan per versi untuk menghitung `jumlah_versi`."},{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Inklusif. Default: tanggal ulasan terawal.","title":"From"},"description":"Inklusif. Default: tanggal ulasan terawal."},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Inklusif. Default: tanggal ulasan terakhir.","title":"To"},"description":"Inklusif. Default: tanggal ulasan terakhir."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KpiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/trend":{"get":{"tags":["agregat"],"summary":"Deret waktu proporsi negatif & volume","operationId":"baca_trend_api_trend_get","parameters":[{"name":"granularity","in":"query","required":false,"schema":{"$ref":"#/components/schemas/Granularity","default":"month"}},{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Inklusif. Default: tanggal ulasan terawal.","title":"From"},"description":"Inklusif. Default: tanggal ulasan terawal."},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Inklusif. Default: tanggal ulasan terakhir.","title":"To"},"description":"Inklusif. Default: tanggal ulasan terakhir."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrendResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/versions":{"get":{"tags":["agregat"],"summary":"Proporsi sentimen per versi minor aplikasi","operationId":"baca_versions_api_versions_get","parameters":[{"name":"min_reviews","in":"query","required":false,"schema":{"type":"integer","minimum":100,"description":"Minimum 100: materialized view agg_version sudah menyaring pada ambang itu (ml/config.yaml versions.min_reviews_per_version), sehingga nilai lebih kecil tidak dapat dilayani tanpa menyesatkan.","title":"Min Reviews"},"description":"Minimum 100: materialized view agg_version sudah menyaring pada ambang itu (ml/config.yaml versions.min_reviews_per_version), sehingga nilai lebih kecil tidak dapat dilayani tanpa menyesatkan."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VersionsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/topics":{"get":{"tags":["agregat"],"summary":"Peringkat kategori keluhan","operationId":"baca_topics_api_topics_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"default":20,"title":"Limit"}},{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"From"}},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"To"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TopicsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/topics/{topic_id}/reviews":{"get":{"tags":["agregat"],"summary":"Drill-down ulasan mentah per kategori","operationId":"baca_topic_reviews_api_topics__topic_id__reviews_get","parameters":[{"name":"topic_id","in":"path","required":true,"schema":{"type":"integer","title":"Topic Id"}},{"name":"sort_by","in":"query","required":false,"schema":{"$ref":"#/components/schemas/SortBy","default":"likes"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"title":"Page Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TopicReviewsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/model/metrics":{"get":{"tags":["model"],"summary":"Perbandingan seluruh model per eval_set","operationId":"baca_model_metrics_api_model_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelMetricsResponse"}}}}}}},"/api/predict":{"post":{"tags":["model"],"summary":"Prediksi sentimen satu teks (demo)","operationId":"prediksi_api_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/health":{"get":{"tags":["ops"],"summary":"Kesiapan DB & artefak model","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}}},"components":{"schemas":{"EvalSetBlock":{"properties":{"eval_set":{"type":"string","title":"Eval Set"},"n_eval":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"N Eval","description":"Dari MANIFEST.json bila tersedia"},"models":{"items":{"$ref":"#/components/schemas/MetricRow"},"type":"array","title":"Models"}},"type":"object","required":["eval_set","models"],"title":"EvalSetBlock"},"Granularity":{"type":"string","enum":["month","week"],"title":"Granularity"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"database":{"type":"string","title":"Database"},"model_dimuat":{"type":"boolean","title":"Model Dimuat"},"model_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model Name"}},"type":"object","required":["status","database","model_dimuat","model_name"],"title":"HealthResponse"},"KpiResponse":{"properties":{"total_ulasan":{"type":"integer","title":"Total Ulasan"},"persen_negatif":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Persen Negatif","description":"Proporsi negatif menurut model produksi"},"baseline_negatif":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Baseline Negatif","description":"Proporsi negatif menurut LABEL AKTUAL (rating 1–2), dihitung atas ulasan berlabel saja — rating 3 tidak punya label biner. Pembanding untuk persen_negatif."},"rata_rating":{"type":"number","maximum":5.0,"minimum":1.0,"title":"Rata Rating"},"jumlah_versi":{"type":"integer","title":"Jumlah Versi","description":"Versi minor yang lolos ambang min_reviews"},"versi_null_pct":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Versi Null Pct"},"model_produksi":{"type":"string","title":"Model Produksi"},"periode":{"$ref":"#/components/schemas/PeriodeInfo"},"catatan":{"items":{"type":"string"},"type":"array","title":"Catatan"}},"type":"object","required":["total_ulasan","persen_negatif","baseline_negatif","rata_rating","jumlah_versi","versi_null_pct","model_produksi","periode","catatan"],"title":"KpiResponse"},"MetricRow":{"properties":{"model_name":{"type":"string","title":"Model Name"},"accuracy":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Accuracy"},"precision_neg":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Precision Neg"},"recall_neg":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Recall Neg"},"f1_neg":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"F1 Neg"},"macro_f1":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Macro F1"},"train_seconds":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Train Seconds"},"is_baseline":{"type":"boolean","title":"Is Baseline"},"is_produksi":{"type":"boolean","title":"Is Produksi"}},"type":"object","required":["model_name","accuracy","precision_neg","recall_neg","f1_neg","macro_f1","train_seconds","is_baseline","is_produksi"],"title":"MetricRow"},"ModelMetricsResponse":{"properties":{"model_produksi":{"type":"string","title":"Model Produksi"},"primary_metric":{"type":"string","title":"Primary Metric"},"majority_baseline_accuracy":{"type":"number","title":"Majority Baseline Accuracy"},"catatan":{"items":{"type":"string"},"type":"array","title":"Catatan"},"eval_sets":{"items":{"$ref":"#/components/schemas/EvalSetBlock"},"type":"array","title":"Eval Sets"}},"type":"object","required":["model_produksi","primary_metric","majority_baseline_accuracy","catatan","eval_sets"],"title":"ModelMetricsResponse"},"PeriodeInfo":{"properties":{"dari":{"type":"string","format":"date","title":"Dari","description":"Batas bawah rentang yang benar-benar dipakai"},"sampai":{"type":"string","format":"date","title":"Sampai","description":"Batas atas rentang yang benar-benar dipakai"},"bulan_parsial":{"items":{"type":"string","format":"date"},"type":"array","title":"Bulan Parsial","description":"Bulan dengan cakupan hari tidak penuh; ikut dihitung, tetapi harus dianotasi di grafik."}},"type":"object","required":["dari","sampai"],"title":"PeriodeInfo"},"PredictRequest":{"properties":{"text":{"type":"string","maxLength":1000,"minLength":1,"title":"Text","example":"aplikasi error terus gabisa bayar"}},"type":"object","required":["text"],"title":"PredictRequest"},"PredictResponse":{"properties":{"text":{"type":"string","title":"Text"},"text_preprocessed":{"type":"string","title":"Text Preprocessed","description":"Masukan TF-IDF: hasil pipeline penuh, sudah ter-stem"},"text_normalized":{"type":"string","title":"Text Normalized","description":"Setelah normalisasi slang, sebelum stopword & stemming — versi yang masih terbaca manusia"},"prediction":{"type":"integer","maximum":1.0,"minimum":0.0,"title":"Prediction"},"label":{"type":"string","title":"Label"},"confidence":{"anyOf":[{"type":"number","maximum":1.0,"minimum":0.0},{"type":"null"}],"title":"Confidence","description":"null bila estimator tidak punya predict_proba"},"model_name":{"type":"string","title":"Model Name"},"note":{"type":"string","title":"Note"}},"type":"object","required":["text","text_preprocessed","text_normalized","prediction","label","model_name","note"],"title":"PredictResponse"},"ReviewItem":{"properties":{"id":{"type":"integer","title":"Id"},"ulasan":{"type":"string","title":"Ulasan","description":"Teks ASLI, bukan ulasan_clean"},"rating":{"type":"integer","maximum":5.0,"minimum":1.0,"title":"Rating"},"tanggal":{"type":"string","format":"date-time","title":"Tanggal"},"likes":{"type":"integer","title":"Likes"},"versi_app":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Versi App"},"pred_label":{"type":"integer","maximum":1.0,"minimum":0.0,"title":"Pred Label"},"pred_proba":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Pred Proba","description":"P(kelas POSITIF) dari pre-scoring batch Fase 6 — bukan keyakinan atas kelas terpilih. pred_label=0 dengan pred_proba=0,06 berarti model 94% yakin ulasan ini negatif."},"sentimen_aktual":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Sentimen Aktual"}},"type":"object","required":["id","ulasan","rating","tanggal","likes","versi_app","pred_label","sentimen_aktual"],"title":"ReviewItem"},"SortBy":{"type":"string","enum":["likes","tanggal","rating"],"title":"SortBy"},"TopicItem":{"properties":{"id":{"type":"integer","title":"Id","description":"Id kategori — min(topics.id) milik kategori itu"},"kategori":{"type":"string","title":"Kategori"},"jumlah_ulasan":{"type":"integer","title":"Jumlah Ulasan"},"jumlah_negatif":{"type":"integer","title":"Jumlah Negatif"},"persen_negatif":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Persen Negatif"},"total_likes":{"type":"integer","title":"Total Likes"},"rata_rating":{"anyOf":[{"type":"number","maximum":5.0,"minimum":1.0},{"type":"null"}],"title":"Rata Rating"}},"type":"object","required":["id","kategori","jumlah_ulasan","jumlah_negatif","persen_negatif","total_likes"],"title":"TopicItem"},"TopicReviewsResponse":{"properties":{"topic_id":{"type":"integer","title":"Topic Id"},"kategori":{"type":"string","title":"Kategori"},"sort_by":{"$ref":"#/components/schemas/SortBy"},"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"},"total_items":{"type":"integer","title":"Total Items"},"total_pages":{"type":"integer","title":"Total Pages"},"data":{"items":{"$ref":"#/components/schemas/ReviewItem"},"type":"array","title":"Data"}},"type":"object","required":["topic_id","kategori","sort_by","page","page_size","total_items","total_pages","data"],"title":"TopicReviewsResponse"},"TopicsResponse":{"properties":{"kategori_ditampilkan":{"type":"integer","title":"Kategori Ditampilkan","description":"Banyak baris pada `data` (dibatasi `limit`)"},"total_kategori":{"type":"integer","title":"Total Kategori","description":"Seluruh kategori yang ada, terlepas dari `limit`"},"total_ulasan_bertopik":{"type":"integer","title":"Total Ulasan Bertopik","description":"Ulasan UNIK yang menyentuh minimal satu kategori. Bukan jumlah dari kolom jumlah_ulasan — satu ulasan dapat masuk beberapa kategori sekaligus."},"model_produksi":{"type":"string","title":"Model Produksi"},"catatan":{"items":{"type":"string"},"type":"array","title":"Catatan"},"data":{"items":{"$ref":"#/components/schemas/TopicItem"},"type":"array","title":"Data"}},"type":"object","required":["kategori_ditampilkan","total_kategori","total_ulasan_bertopik","model_produksi","catatan","data"],"title":"TopicsResponse"},"TrendPoint":{"properties":{"periode":{"type":"string","format":"date","title":"Periode"},"total_ulasan":{"type":"integer","title":"Total Ulasan"},"jumlah_negatif":{"type":"integer","title":"Jumlah Negatif"},"persen_negatif":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Persen Negatif"},"rata_rating":{"anyOf":[{"type":"number","maximum":5.0,"minimum":1.0},{"type":"null"}],"title":"Rata Rating"},"total_likes":{"type":"integer","title":"Total Likes"},"periode_parsial":{"type":"boolean","title":"Periode Parsial"}},"type":"object","required":["periode","total_ulasan","jumlah_negatif","persen_negatif","total_likes","periode_parsial"],"title":"TrendPoint"},"TrendResponse":{"properties":{"granularity":{"$ref":"#/components/schemas/Granularity"},"model_produksi":{"type":"string","title":"Model Produksi"},"catatan":{"items":{"type":"string"},"type":"array","title":"Catatan"},"data":{"items":{"$ref":"#/components/schemas/TrendPoint"},"type":"array","title":"Data"}},"type":"object","required":["granularity","model_produksi","catatan","data"],"title":"TrendResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VersionItem":{"properties":{"versi_minor":{"type":"string","title":"Versi Minor"},"total_ulasan":{"type":"integer","title":"Total Ulasan"},"jumlah_negatif":{"type":"integer","title":"Jumlah Negatif"},"persen_negatif":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Persen Negatif"},"rata_rating":{"anyOf":[{"type":"number","maximum":5.0,"minimum":1.0},{"type":"null"}],"title":"Rata Rating"},"pertama_muncul":{"type":"string","format":"date-time","title":"Pertama Muncul"}},"type":"object","required":["versi_minor","total_ulasan","jumlah_negatif","persen_negatif","pertama_muncul"],"title":"VersionItem"},"VersionsResponse":{"properties":{"min_reviews_threshold":{"type":"integer","title":"Min Reviews Threshold"},"versions_included":{"type":"integer","title":"Versions Included"},"coverage_pct":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Coverage Pct","description":"Ulasan yang tercakup versi lolos-ambang, sebagai persen dari ulasan yang PUNYA versi_minor."},"coverage_pct_seluruh_ulasan":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Coverage Pct Seluruh Ulasan","description":"Persentase yang sama, tetapi atas seluruh 100.000 ulasan — termasuk yang versi_app-nya NULL."},"null_version_pct":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Null Version Pct"},"model_produksi":{"type":"string","title":"Model Produksi"},"catatan":{"items":{"type":"string"},"type":"array","title":"Catatan"},"data":{"items":{"$ref":"#/components/schemas/VersionItem"},"type":"array","title":"Data"}},"type":"object","required":["min_reviews_threshold","versions_included","coverage_pct","coverage_pct_seluruh_ulasan","null_version_pct","model_produksi","catatan","data"],"title":"VersionsResponse"}}}}